addition of testcases
