--- test/.coveragerc
@@ -1,39 +1,0 @@
-# .coveragerc to control coverage.py
-[run]
-branch = True
-omit =
-    */.circleci/*
-    /node_modules/*
-    /sql_script/*
-    /venv_mplace/*
-    Crypto/*
-    chardet/*
-    eth_utils/*
-    pycparser/*
-    pymysql/*
-    rlp/*
-    requests/*
-    toolz/*
-    urllib3/*
-    websockets/*
-    /usr/*
-    web3/*
-    parsimonious/*
-    idna/*
-    eth_rlp/*
-    hexbytes/*
-    ens/*
-    cytoolz/*
-    cryptography/*
-    cffi/*
-    certifi/*
-    attrdict/*
-    eth_typing/*
-    eth_keys/*
-    eth_hash/*
-    eth_keyfile/*
-    eth_account/*
-    eth_abi/*
-    six.py
-    schema.py
-    contract_api/search.py
